Add interref role, a in-org alt to intersphinx
Interref role suffixes the ref role with the format ref-<repo> (e.g.
ref-hdl) to create cross references between different docs inside the
org.

Currently, the tracked repos are manually set on lut.py, with plans to
automate or self-service the inclusion of new repositories.

For versioning handling and multiple hosted, the environment variables
are used:
* ADOC_INTERREF_URI: set the host uri, can be a link or local.
* ADOC_INTERREF_TAG: enable tag suffix, e.g. pyadi-iio/main.
* ADOC_INTERREF_RELEASE: enable release tag suffix, e.g. pyadi-iio/v1.0.0

Signed-off-by: Jorge Marques <jorge.marques@analog.com>