Simplify filtered tree logic, bump version
Always hide other toctrees when on a page included on a toctree.
Do not rely on doctools lut.py anymore, since was confusing having to
update it to affect other repository.
The "topic" entry on the lut.py is now the path to the index.rst page
of the toctree, generally an orphan page, to reduce nesting by one level.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>