directives: hdl: revert d70ba6a93c5f7a47e06fc599e705187feef303c8
Revert media_print for hdl directives.
The reason is that improvements to the CSS fixes the original need
to have a different output for PDFs.
The media_print variable should be used in scenarios where it is
explicitly needed to compile the content in a different manner than for
the hosted html.
For example, to render some content that would instead be rendered with
JavaScript libraries in the user browser.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>