lut: change pathname to conf.py dir
Don't parse the Makefile anymore to obtain the builddir and sourcedir,
instead, pathname is already the sourcedir (where conf.py is at) and
builddir is the default '_build'.
The motivation is to simplify things and I have seen horribles things
done to the default sphinx makefile, so the tools shouldn't arbitrarily
run that.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>