custom-doc: add git lfs smudge
Allows to fetch-on-time git lfs artifacts, reducing the burden of
downloading all lfs artifacts of the current checkout.
Only images from "Copying images..." are considered.
Downloadable files should also be looked into, but since the focus is
PDF, the solution for downloadable files may be to actually resolve to a
web link.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>