scripts: always fetch extra css/js
Reduce page loading time by always fetching the extra css and js,
instead of waiting the metadata.json to tell so.
Also, use state to infer from where to fetch, instead of trying
multiple options.
For backwards-compatibility, check if app contains fetch, if so, we can
safely call fetch.then, the method that collects callbacks to call when
metadata.json is loaded. If not, that means the fetch version will
append extra js after metadata.json loads, so no need to have a
callback.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>